CORE-523 public sharing

+++ web-app/js/unifina/sharing-dialog/sharing-dialog.js
@@ -115,6 +115,8 @@
         '<div class="owner-row">' +
             '<span>Owner: </span>' +
             '<span class="owner-label"><%= owner %></span>' +
+            '<input type="checkbox" class="anonymous-switcher" <%= checked ? "checked" : "" %>>' +
+            '<div class="publish-label"> Allow anonymous read access </div>' +
         '</div>' +
         '<div class="access-list"></div>' +
         '<div class="form-inline new-user-row">' +
@@ -124,6 +126,34 @@
     )
 
     var AccessListView = Backbone.View.extend({
+        initialize: function(args) {
+            this.$el.html(accessListTemplate({
+                owner: args.owner,
+                checked: !!originalAnonPermission
+            }))
+            this.$ownerLabel = this.$(".owner-label")
+            this.$accessList = this.$(".access-list")
+            this.$newUserField = this.$(".new-user-field")
+            this.$anonymousCheckbox = this.$(".anonymous-switcher")
+
+            this.$anonymousCheckbox.switcher({
+                on_state_content: '<i class="fa fa-check"></i>',
+                off_state_content: '<i class="fa fa-times"></i>',
+            })
+            this.$(".switcher-toggler").html('<i class="fa fa-globe">')
+
+            this.listenTo(accessList, 'add', this.createAccessView)
+            this.listenTo(accessList, 'reset', this.addAll)
+        },
+        createAccessView: function(accessRow) {
+            var view = new AccessView({model: accessRow})
+            this.$accessList.append(view.render().$el)
+        },
+        addAll: function() {
+            //this.$accessList.html()
+            accessList.each(this.createAccessView, this)
+        },
+
         events: {
             "click .new-user-button": "finishUserInput",
             "keyup .new-user-field": "keyHandler"
@@ -159,30 +189,6 @@
             })
             this.$newUserField.val("")
             return "added"
-        },
-
-        initialize: function(args) {
-            this.$el.html(accessListTemplate({
-                owner: args.owner
-            }))
-            this.$ownerLabel = this.$(".owner-label")
-            this.$accessList = this.$(".access-list")
-            this.$newUserField = this.$(".new-user-field")
-
-            this.listenTo(accessList, 'add', this.createAccessView)
-            this.listenTo(accessList, 'reset', this.addAll)
-        },
-
-        accessViews: [],
-        createAccessView: function(accessRow) {
-            var view = new AccessView({model: accessRow})
-            this.$accessList.append(view.render().$el)
-            this.accessViews.push(view)
-        },
-
-        addAll: function() {
-            //this.$accessList.html()
-            accessList.each(this.createAccessView, this)
         }
     })
     
@@ -218,6 +224,7 @@
     var resourceUrl;
     var originalPermissions = {};   // { username: { operation: Permission object } }
     var originalOwner;
+    var originalAnonPermission;
 
     /** Entry point */
     exports.sharePopup = function(url, resourceName) {
@@ -254,28 +261,33 @@
             })
             $(".resource-name-label").text(resourceName || urlToResourceName(resourceUrl))
 
-            // TODO: find out who is stealing the focus, then remove this kludge...
-            setTimeout(function() { $(".new-user-field").focus() }, 500)
+            // map list of Permissions to list of Access rows on the form (combine)
+            originalPermissions = {}
+            var initialAccessMap = {}
+            originalAnonPermission = undefined;
+            _(originalPermissionList).each(function(p) {
+                if (!p || !p.operation) { return }      // continue
+                if (p.anonymous) {
+                    originalAnonPermission = p;
+                } else if (p.user) {
+                    if (!originalPermissions[p.user]) {
+                        originalPermissions[p.user] = {}
+                        initialAccessMap[p.user] = {user: p.user}
+                    }
+                    originalPermissions[p.user][p.operation] = p
+                    initialAccessMap[p.user][p.operation] = true
+                }
+            })
 
             listView = new AccessListView({
                 el: ".bootbox .modal-body",
                 id: "access-list",
                 owner: originalOwner
             })
-
-            // map list of Permissions to list of Access rows on the form (combine)
-            originalPermissions = {}
-            var initialAccessMap = {}
-            _(originalPermissionList).each(function(p) {
-                if (!p || !p.user || !p.operation) { return }   // continue
-                if (!originalPermissions[p.user]) {
-                    originalPermissions[p.user] = {}
-                    initialAccessMap[p.user] = { user: p.user }
-                }
-                originalPermissions[p.user][p.operation] = p
-                initialAccessMap[p.user][p.operation] = true
-            })
             accessList.reset(_.values(initialAccessMap))
+
+            // TODO: find out who is stealing the focus, then remove this kludge...
+            setTimeout(function() { $(".new-user-field").focus() }, 500)
         })
 
         return false
@@ -298,10 +310,6 @@
             return false;
         }
 
-        _(listView.accessViews).each(function(view) {
-            view.$(".user-access-row").removeClass("has-error")
-        })
-
         // find changes
         var addedPermissions = []
         var removedPermissions = []
@@ -310,9 +318,9 @@
             var user = accessModel.attributes.user
             var before = originalPermissions[user] || {}
             var after = accessModel.attributes
-            if (before.read  && !after.read)  { removedPermissions.push(before.read.id) }
-            if (before.write && !after.write) { removedPermissions.push(before.write.id) }
-            if (before.share && !after.share) { removedPermissions.push(before.share.id) }
+            if (before.read  && !after.read)  { removedPermissions.push(before.read) }
+            if (before.write && !after.write) { removedPermissions.push(before.write) }
+            if (before.share && !after.share) { removedPermissions.push(before.share) }
             if (!before.share && after.share) { addedPermissions.push({user: user, operation: "share"}) }
             if (!before.write && after.write) { addedPermissions.push({user: user, operation: "write"}) }
             if (!before.read  && after.read)  { addedPermissions.push({user: user, operation: "read"}) }
@@ -321,13 +329,12 @@
         // completely removed users don't show up in accessList, need to be tested separately
         _(originalPermissions).each(function (before, user) {
             if (user in testedUsers) { return }    // continue
-            if (before.read)  { removedPermissions.push(before.read.id) }
-            if (before.write) { removedPermissions.push(before.write.id) }
-            if (before.share) { removedPermissions.push(before.share.id) }
+            if (before.read)  { removedPermissions.push(before.read) }
+            if (before.write) { removedPermissions.push(before.write) }
+            if (before.share) { removedPermissions.push(before.share) }
         })
 
-        // close modal directly if no changes
-        if (addedPermissions.length == 0 && removedPermissions.length == 0) { return true; }
+        var isAnon = listView.$anonymousCheckbox.prop("checked")
 
         // generate one API call for each change
         var started = 0
@@ -336,39 +343,70 @@
         var grantedTo = {}
         var revokedFrom = {}
         var errorMessages = []
-        _(addedPermissions).each(function(permission) {
+        if (addedPermissions.length > 0 || removedPermissions.length > 0) {
+            _(addedPermissions).each(function(permission) {
+                started += 1
+                pendingRequests.push($.ajax({
+                    url: resourceUrl + "/permissions",
+                    method: "POST",
+                    data: JSON.stringify(permission),
+                    contentType: "application/json",
+                    error: onError,
+                    success: function(response) {
+                        grantedTo[response.user] = true
+                        // update state so that the next diff goes correctly
+                        if (!originalPermissions[response.user]) { originalPermissions[response.user] = {} }
+                        originalPermissions[response.user][response.operation] = response
+                        onSuccess()
+                    }
+                }))
+            })
+            _(removedPermissions).each(function(p) {
+                started += 1
+                pendingRequests.push($.ajax({
+                    url: resourceUrl + "/permissions/" + p.id,
+                    method: "DELETE",
+                    contentType: "application/json",
+                    error: onError,
+                    success: function() {
+                        revokedFrom[p.user] = true
+                        // update state so that the next diff goes correctly
+                        delete originalPermissions[p.user][p.operation]
+                        onSuccess()
+                    }
+                }))
+            })
+        }
+        if (isAnon && !originalAnonPermission) {
             started += 1
             pendingRequests.push($.ajax({
                 url: resourceUrl + "/permissions",
                 method: "POST",
-                data: JSON.stringify(permission),
+                data: JSON.stringify({anonymous: true, operation: "read"}),
                 contentType: "application/json",
                 error: onError,
                 success: function(response) {
-                    grantedTo[response.user] = true
-                    // update state so that the next diff goes correctly
-                    if (!originalPermissions[response.user]) { originalPermissions[response.user] = {} }
-                    originalPermissions[response.user][response.operation] = response
+                    grantedTo["the whole world"] = true
+                    originalAnonPermission = response
                     onSuccess(response)
                 }
             }))
-        })
-        _(removedPermissions).each(function(id) {
+        }
+        if (!isAnon && !!originalAnonPermission) {
             started += 1
             pendingRequests.push($.ajax({
-                url: resourceUrl + "/permissions/" + id,
+                url: resourceUrl + "/permissions/" + originalAnonPermission.id,
                 method: "DELETE",
                 contentType: "application/json",
                 error: onError,
-                success: function(response) {
-                    revokedFrom[response.user] = true
-                    // update state so that the next diff goes correctly
-                    delete originalPermissions[response.user][response.operation]
-                    onSuccess(response)
+                success: function() {
+                    revokedFrom["anonymous users"] = true
+                    originalAnonPermission = undefined
+                    onSuccess()
                 }
             }))
-        })
-        function onSuccess(response) {
+        }
+        function onSuccess() {
             successful += 1
             if (successful + failed === started) { onDone() }
         }
@@ -395,8 +433,8 @@
             }
         }
 
-        // don't close the dialog yet, saving is still in progress...
-        return false
+        // don't close the dialog yet if there was something to save
+        return started == 0
     }
 
     exports.sharePopup.closeAndSaveChanges = function() {
